refactor : 로또 번호 new 로 생성할 수 있도록 수정, WinningResult 기능 WinningNumber 객체로 이동
