+++ src/main/java/lotto/domain/lotto/LottoNumber.java
@@ -1,36 +1,34 @@
 package lotto.domain.lotto;
 
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.Objects;
 
 public class LottoNumber {
     public static final int MIN_LOTTO_BOUND = 1;
     public static final int MAX_LOTTO_BOUND = 45;
-    private static final Map<Integer, LottoNumber> LOTTO_NUMBER_BOX = generate();
 
     private final int lottoNumber;
 
-    private LottoNumber(int lottoNumber) {
+    public LottoNumber(int lottoNumber) {
+        validateRange(lottoNumber);
         this.lottoNumber = lottoNumber;
     }
 
-    public static LottoNumber from(int number) {
-        Optional<LottoNumber> lottoNumber = Optional.ofNullable(LOTTO_NUMBER_BOX.get(number));
-        return lottoNumber.orElseThrow(() -> new IllegalArgumentException("로또 번호는 1 ~ 45 까지 입니다."));
+    private static void validateRange(int number) {
+        if (number < 1 || number > 45) {
+            throw new IllegalArgumentException("로또 번호는 1 ~ 45 까지 입니다.");
+        }
     }
 
-    private static Map<Integer, LottoNumber> generate() {
-        return IntStream.rangeClosed(MIN_LOTTO_BOUND, MAX_LOTTO_BOUND)
-                .mapToObj(LottoNumber::new)
-                .collect(Collectors.toMap(
-                        LottoNumber::getLottoNumber, Function.identity()
-                ));
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        LottoNumber that = (LottoNumber) o;
+        return lottoNumber == that.lottoNumber;
     }
 
-    public int getLottoNumber() {
-        return lottoNumber;
+    @Override
+    public int hashCode() {
+        return Objects.hash(lottoNumber);
     }
 }
