+++ src/test/java/lotto/domain/lotto/LottoNumberTest.java
@@ -15,10 +15,10 @@ class LottoNumberTest {
     @DisplayName("1 ~ 45 사이의 값으로 로또 번호 객체를 생성한다.")
     void createLottoNumber() {
         //given
-        LottoNumber lottoNumber = LottoNumber.from(1);
+        LottoNumber lottoNumber = new LottoNumber(1);
 
         //when //then
-        assertThat(lottoNumber).isEqualTo(LottoNumber.from(1));
+        assertThat(lottoNumber).isEqualTo(new LottoNumber(1));
     }
 
     @ParameterizedTest(name = "로또 숫자 범위: [{index}] {0}")
@@ -26,7 +26,7 @@ class LottoNumberTest {
     @DisplayName("로또 번호가 1 ~ 45가 아닐 경우 예외가 발생한다.")
     void validateRange(int invalidLottoNumber) {
         //given //when //then
-        assertThatThrownBy(() -> LottoNumber.from(invalidLottoNumber))
+        assertThatThrownBy(() -> new LottoNumber(invalidLottoNumber))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
